Move sync progress logic to history manager
